fixed bug in packetReceive()
This fixes the bug described in issue #182. If gNextPacketPtr=0 we
have been writing 0xFFFF into ERXRDPT (the register in the enc28j60
that protects the existing content of the receive buffer to be
overwritten by new packets). As a result if in this state lots of
new packets arrive the whole buffer turned into garbage and
the receive logic would stop working. This bug showed up albeit only
rarely and in high load scenarios.

I checked with other libraries and the datasheet, and this is the
way that ERXRDPT is usually updated.