add ifdefs to reduce functionality
The linker cannot efficiently remove unused code because a lot of
stuff may be called from packetLoop(). This commit adds ifdefs that
allow to easily remove some of these calls. For example, when not
using DHCP, one can set the corresponding #define (in EtherCard.h)
to 0 and the code size reduces by 1550 byte.

This solution is not really ideal since in some projects one might
need DHCP in others not. The alternative would be to add parameters
to packetLoop(). Then a call like packetLoop(len, 1) would, e.g.,
mean that the packetLoop should use DHCP and packetLoop(len, 0) means
the packetLoop should not call DHCP. In theory, if using link time
optimization, the compiler should be able to optimize any DHCP related
stuff away when using the second form. However, this does not seem to
be working. The compiler optimizes something away but not all of it.