add readPacketSlice() to read a packet that is larger than buffer
This function allows to read the payload of a packet in several slices,
that each fit into the global buffer. For this to work the receive
function has been changed to not free the memory of the current packet
within the receive buffer. This memory is freed in the next call to
packetReceive(), instead.