Removed references to MACON2 (#317)
With thanks to @mindstormsking:
The MACON2 register I have removed does not exist. The datasheet of the ENC28J60 never mentions it or any of its bits, but states its address as Reserved. According to the datasheet "All reserved registers may be read, but their contents must not be changed.", so a write operation should not be performed.